handler fixes, moer debug

+++ signals.cpp
@@ -19,8 +19,12 @@ void ctrlZHandler(int sig_num) { //todo: check with pipe and redirections
         else{
             JobsList::JobEntry* pipeJob1 = smash.jobsList->getJobByPid(pipePid1);
             pipeJob1->state = STOPPED;
-            smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID);
-            smash.jobsList->maxJobID++;
+            if (pipeJob1->formerJobId == NOVALUE){
+                smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID);
+                smash.jobsList->maxJobID++;
+            } else {
+                smash.jobsList->changeJobId(pipeJob1, pipeJob1->formerJobId);
+            }
             cout << "smash: process " << pipePid1<< " was stopped" << endl;
         }
         if (kill(pipePid2, SIGSTOP) != 0){
@@ -30,26 +34,41 @@ void ctrlZHandler(int sig_num) { //todo: check with pipe and redirections
         else{
             JobsList::JobEntry* pipeJob2 = smash.jobsList->getJobByPid(pipePid2);
             pipeJob2->state = STOPPED;
-            smash.jobsList->changeJobId(pipeJob2, smash.jobsList->maxJobID);
-            smash.jobsList->maxJobID++;
+            if (pipeJob2->formerJobId == NOVALUE){
+                smash.jobsList->changeJobId(pipeJob2, smash.jobsList->maxJobID);
+                smash.jobsList->maxJobID++;
+            } else {
+                smash.jobsList->changeJobId(pipeJob2, pipeJob2->formerJobId);
+            }
+
             cout << "smash: process " << pipePid2 << " was stopped" << endl;
         }
         return;
     }
     JobsList::JobEntry* fgJob = smash.jobsList->getFgJob();
+    if (fgJob == NULL){
+        cerr << "smash error: there is no foreground job to stop" << endl;
+        return;
+    }
+//    cout << "fgJob jobID in ctrl z" << fgJob->jobID << endl; // todo debug
+//    cout << "fgJob PID in ctrl z" << fgJob->command->getPID() << endl; // todo debug
     if (kill(fgJob->command->getPID(), SIGSTOP) != 0){
         perror("smash error: kill failed"); // kill is the syscall that failed
         return;
     }
     fgJob->state = STOPPED;
-    smash.jobsList->changeJobId(fgJob, smash.jobsList->maxJobID);
-    smash.jobsList->maxJobID++;
+    if (fgJob->formerJobId == NOVALUE){
+        smash.jobsList->changeJobId(fgJob, smash.jobsList->maxJobID);
+        smash.jobsList->maxJobID++;
+    } else {
+        smash.jobsList->changeJobId(fgJob, fgJob->formerJobId);
+    }
     cout << "smash: process " << fgJob->command->getPID() << " was stopped" << endl;
     return;
 }
 
 void ctrlCHandler(int sig_num) { //todo: check with pipe and redirections
-    cout << "smash got ctrl-c" << endl;
+    cout << "smash: got ctrl-c" << endl;
     SmallShell& smash = SmallShell::getInstance();
     if (smash.jobsList->hasPipeInFg){
         pid_t pipePid1 = smash.jobsList->pipePid1;
@@ -73,6 +92,10 @@ void ctrlCHandler(int sig_num) { //todo: check with pipe and redirections
         return;
     }
     JobsList::JobEntry* fgJob = smash.jobsList->getFgJob();
+    if (fgJob == NULL){
+        cerr << "smash error: there is no foreground job to kill" << endl;
+        return;
+    }
     if (kill(fgJob->command->getPID(), SIGKILL) != 0){
         perror("smash error: kill failed"); // kill is the syscall that failed
         return;
